# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClaimLens API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Get Health",
        "description": "Liveness probe.",
        "operationId": "get_health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          }
        }
      }
    },
    "/api/config": {
      "get": {
        "summary": "Get Config",
        "description": "Public config: provider and model names only, never keys.",
        "operationId": "get_config_api_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs": {
      "post": {
        "summary": "Create Run",
        "description": "Accept a paper PDF and start the pipeline in a background worker.",
        "operationId": "create_run_api_runs_post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_run_api_runs_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Runs",
        "description": "List runs, oldest first. ``?demo=true`` shows only demo runs.",
        "operationId": "get_runs_api_runs_get",
        "parameters": [
          {
            "name": "demo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Demo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  },
                  "title": "Response Get Runs Api Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "summary": "Get Run",
        "description": "Run detail with per-stage status, timestamps and errors.",
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/events": {
      "get": {
        "summary": "Stream Run Events",
        "description": "SSE stream: replay ``events.jsonl``, then stream live events.",
        "operationId": "stream_run_events_api_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/code": {
      "get": {
        "summary": "Get Run Code",
        "description": "File tree of generated code per claim and iteration (server-enumerated).",
        "operationId": "get_run_code_api_runs__run_id__code_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeTree"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/code/{claim_id}/{iteration}/{file_index}": {
      "get": {
        "summary": "Get Run Code File",
        "description": "File content from the server-enumerated code tree.\n\nFiles resolve only from the enumeration above, never from a raw\npath, so ``..`` segments and unknown names return 404.",
        "operationId": "get_run_code_file_api_runs__run_id__code__claim_id___iteration___file_index__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "claim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Claim Id"
            }
          },
          {
            "name": "iteration",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Iteration"
            }
          },
          {
            "name": "file_index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "File Index"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/{artifact}": {
      "get": {
        "summary": "Get Run Artifact",
        "description": "Fetch one run artifact. Names are whitelisted; paths are fixed.",
        "operationId": "get_run_artifact_api_runs__run_id___artifact__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "artifact",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/cancel": {
      "post": {
        "summary": "Cancel Run",
        "description": "Cancel a queued or running run.",
        "operationId": "cancel_run_api_runs__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_create_run_api_runs_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_run_api_runs_post"
      },
      "CodeClaim": {
        "properties": {
          "claim_id": {
            "type": "string",
            "title": "Claim Id"
          },
          "iterations": {
            "items": {
              "$ref": "#/components/schemas/CodeIteration"
            },
            "type": "array",
            "title": "Iterations",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "claim_id"
        ],
        "title": "CodeClaim",
        "description": "Generated code for one claim across iterations."
      },
      "CodeFile": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "index",
          "name",
          "size"
        ],
        "title": "CodeFile",
        "description": "One generated code file, addressed by server-assigned index."
      },
      "CodeIteration": {
        "properties": {
          "iteration": {
            "type": "integer",
            "title": "Iteration"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/CodeFile"
            },
            "type": "array",
            "title": "Files",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "iteration"
        ],
        "title": "CodeIteration",
        "description": "One saved iteration of a claim's generated code."
      },
      "CodeTree": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/CodeClaim"
            },
            "type": "array",
            "title": "Claims",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "CodeTree",
        "description": "File tree of generated code, enumerated by the server."
      },
      "ConfigInfo": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model_planner": {
            "type": "string",
            "title": "Model Planner"
          },
          "model_agent": {
            "type": "string",
            "title": "Model Agent"
          },
          "model_fast": {
            "type": "string",
            "title": "Model Fast"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "model_planner",
          "model_agent",
          "model_fast"
        ],
        "title": "ConfigInfo",
        "description": "Public runtime config. Never includes API keys."
      },
      "CreateRunResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          }
        },
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "CreateRunResponse",
        "description": "Ids for a newly accepted run."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Health": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "Health",
        "description": "Liveness response."
      },
      "RunDetail": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "demo": {
            "type": "boolean",
            "title": "Demo",
            "default": false
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "default": ""
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "default": ""
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "default": ""
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "stages": {
            "additionalProperties": {
              "$ref": "#/components/schemas/StageState"
            },
            "type": "object",
            "title": "Stages",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status"
        ],
        "title": "RunDetail",
        "description": "Full run state including per-stage status."
      },
      "RunSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "demo": {
            "type": "boolean",
            "title": "Demo",
            "default": false
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "default": ""
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "default": ""
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status"
        ],
        "title": "RunSummary",
        "description": "One row of the run list."
      },
      "StageState": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StageState",
        "description": "Status and timestamps for one pipeline stage."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}